+++ ingeniamotion/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.5.2'
+__version__ = '0.5.3'
 """ str: Library version. """
 from .motion_controller import MotionController
 from . import enums