+++ ingeniamotion/enums/__init__.py
@@ -1,4 +1,6 @@
 from enum import IntEnum
+from ingenialink.register import REG_DTYPE, REG_ACCESS
+from ingenialink.canopen import CAN_BAUDRATE, CAN_DEVICE
 
 
 class OperationMode(IntEnum):
@@ -122,3 +124,15 @@ class MonitoringVersion(IntEnum):
     """Monitoring V2 used for Capitan and some custom low-power drivers."""
     MONITORING_V3 = 2
     """Monitoring V3 used for Everest and Capitan newer than 1.8.1."""
+
+
+class SeverityLevel(IntEnum):
+    """Test result enum"""
+    SUCCESS = 0
+    WARNING = 1
+    FAIL = 2
+
+
+enums = list(globals().keys())
+enums.remove("IntEnum")
+__all__ = enums