+++ ingeniamotion/exceptions.py
@@ -21,3 +21,13 @@ class IMStatusWordError(IMException):
 class IMRegisterNotExist(IMException):
     """Error raised by IngeniaMotion when a register not exists"""
     pass
+
+
+class IMRegisterWrongAccess(IMException):
+    """Error raised by IngeniaMotion when trying to write to a read-only register"""
+    pass
+
+
+class IMTimeoutError(IMException):
+    """Error raised by IngeniaMotion when a timeout has occurred"""
+    pass