+++ tests/test_motion.py
@@ -5,6 +5,7 @@ from ingenialink import exceptions
 
 from ingeniamotion.enums import OperationMode
 from ingeniamotion.motion import Motion
+from ingeniamotion.exceptions import IMTimeoutError
 
 POS_PID_KP_VALUE = 0.1
 POSITION_PERCENTAGE_ERROR_ALLOWED = 5
@@ -314,7 +315,8 @@ def test_wait_for_position_timeout(motion_controller):
     timeout_value = 2
     mc, alias = motion_controller
     init_time = time.time()
-    mc.motion.wait_for_position(10000, servo=alias, timeout=timeout_value)
+    with pytest.raises(IMTimeoutError):
+        mc.motion.wait_for_position(10000, servo=alias, timeout=timeout_value)
     final_time = time.time()
     assert pytest.approx(final_time-init_time, abs=0.1) == timeout_value
 
@@ -324,7 +326,8 @@ def test_wait_for_velocity_timeout(motion_controller):
     timeout_value = 2
     mc, alias = motion_controller
     init_time = time.time()
-    mc.motion.wait_for_velocity(10000, servo=alias, timeout=timeout_value)
+    with pytest.raises(IMTimeoutError):
+        mc.motion.wait_for_velocity(10000, servo=alias, timeout=timeout_value)
     final_time = time.time()
     assert pytest.approx(final_time-init_time, abs=0.1) == timeout_value